+++ livecomponents/management/commands/_templates.py
@@ -1,5 +1,4 @@
-COMPONENT_HTML_TEMPLATE = """
-{{% load livecomponents %}}
+COMPONENT_HTML_TEMPLATE = """{{% load livecomponents %}}
 
 <div {{% component_attrs component_id %}}>
     {component_name} component
@@ -7,32 +6,32 @@ COMPONENT_HTML_TEMPLATE = """
 """
 
 
-COMPONENT_PYTHON_TEMPLATE = """
-from typing import Any
+COMPONENT_PYTHON_TEMPLATE = """from typing import Any
 from django_components import component
-from pydantic import BaseModel
 
 from livecomponents import CallContext, InitStateContext, LiveComponent, command
+from livecomponents.utils import LiveComponentsModel
 
 
-class {component_name}State(BaseModel):
+class {class_name}State(LiveComponentsModel):
+    \"\"\"Define the state of your component here.\"\"\"
     pass
 
 
-@component.register("{lowercase_component_name}")
-class {component_name}(LiveComponent[{component_name}State]):
-    template_name = "{lowercase_component_name}/{lowercase_component_name}.html"
+@component.register("{component_name}")
+class {class_name}Component(LiveComponent[{class_name}State]):
+    template_name = "{component_name}/{proper_name}.html"
 
-    def get_extra_context_data(self, state: {component_name}State) -> dict[str, Any]:
+    def get_extra_context_data(self, state: {class_name}State) -> dict[str, Any]:
         return {{}}
 
     def init_state(
         self, context: InitStateContext, **component_kwargs
-    ) -> {component_name}State:
-        return {component_name}State(**component_kwargs)
+    ) -> {class_name}State:
+        return {class_name}State(**component_kwargs)
 
     # @command
-    # def say_hello(self, call_context: CallContext[{component_name}State]):
+    # def say_hello(self, call_context: CallContext[{class_name}State]):
     #     \"\"\"Example command.\"\"\"
     #     pass
 """