+++ livecomponents/management/commands/createlivecomponent.py
@@ -19,34 +19,70 @@ class Command(BaseCommand):
         parser.add_argument(
             "component_name",
             type=str,
-            help="A camel case name for the component. E.g., 'ClickCounter'",
+            help=(
+                "Component name in snake case. The name can use directory separator "
+                "for creating namespaces. Normally, the first namespace is the "
+                "app name. E.g., 'counters/click_counter'"
+            ),
+        )
+        parser.add_argument(
+            "--class-name",
+            type=str,
+            help=(
+                "Optional class name for the component. "
+                "E.g., 'ClickCounter'. If not given, the component name, "
+                "converted to PascalCase, is used."
+            ),
+        )
+        parser.add_argument(
+            "-f",
+            "--force",
+            action="store_true",
+            help="Overwrite existing files",
         )
 
     def handle(self, *args, **options):
         app_name = options["app_name"]
         component_name = options["component_name"]
+        proper_name = component_name.split("/")[-1]
+
+        class_name = options["class_name"]
+        if class_name is None:
+            class_name = generate_class_name(proper_name)
+        module_name = (
+            f"{app_name}.components.{component_name.replace('/', '.')}.{proper_name}"
+        )
 
         self.stdout.write(
             f"Creating a new live component {component_name!r} in app {app_name!r}"
         )
 
         app_path = self.get_app_path(app_name)
-        component_path = app_path / "components" / component_name.lower()
+        component_path = app_path / "components" / component_name
         component_path.mkdir(parents=True, exist_ok=True)
-        init_py = component_path / "__init__.py"
-        component_py = component_path / f"{component_name.lower()}.py"
-        component_html = component_path / f"{component_name.lower()}.html"
 
-        if component_py.exists():
+        # Create __init__.py files for all the parent directories of the component
+        # file down to "components" directory.
+        parent_path = component_path.parent
+        while parent_path != app_path / "components":
+            init_py = parent_path / "__init__.py"
+            if not init_py.exists():
+                init_py.touch()
+            parent_path = parent_path.parent
+
+        component_py = component_path / f"{proper_name}.py"
+        component_html = component_path / f"{proper_name}.html"
+
+        if component_py.exists() and not options["force"]:
             raise CommandError(f"Component file already exists: {component_py}")
-        if component_html.exists():
+        if component_html.exists() and not options["force"]:
             raise CommandError(f"Component file already exists: {component_html}")
 
         context = {
             "component_name": component_name,
-            "lowercase_component_name": component_name.lower(),
+            "proper_name": proper_name,
+            "class_name": class_name,
         }
-        init_py.touch()
         component_py.write_text(COMPONENT_PYTHON_TEMPLATE.format(**context))
         component_html.write_text(COMPONENT_HTML_TEMPLATE.format(**context))
         self.stdout.write(
@@ -54,9 +90,21 @@ class Command(BaseCommand):
                 f"Created component files: {component_py} and {component_html}"
             )
         )
+        self.stdout.write(
+            f"- If necessary, add {module_name!r} "
+            f"to COMPONENTS['libraries'] of your settings"
+        )
+        self.stdout.write(
+            f'- Use component as {{% livecomponent "{component_name}" %}}'
+        )
 
     def get_app_path(self, app_name: str) -> Path:
         for app_config in apps.get_app_configs():
             if app_config.name == app_name:
                 return Path(app_config.path)
         raise CommandError(f"App not found: {app_name}")
+
+
+def generate_class_name(proper_name: str) -> str:
+    """Generate a class name from the component name."""
+    return "".join(word.capitalize() for word in proper_name.split("_"))