+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+- Made it possible to use nested components. The change required updating the hierarchy separator from "/" to "|" and type separator from "." to ":".
+- Updated the "createlivecomponent" command. The new command accepts the component path and deduces the component name from it.
+
 ## 1.4.2 (2023-11-10)
 
 - Marked LiveComponent as abstract base class.